feat: Add IngredientService

+++ backend/DataAccess/IRepository.cs
@@ -3,9 +3,9 @@
     public interface IRepository<DBEntity, ModelEntity>
     {
         Task<IEnumerable<ModelEntity>> Get(string includeTables = "");
-        Task<ModelEntity> GetOne(long id);
+        Task<ModelEntity> GetOne(int id);
         Task<ModelEntity> Insert(ModelEntity entity);
         Task<ModelEntity> Update(ModelEntity entity);
-        Task<bool> Delete(long id);
+        Task<bool> Delete(int id);
     }
 }
